Minor BlockMover cleanup
