+++ core/src/main/java/nl/pim16aap2/bigDoors/Commander.java
@@ -487,7 +487,7 @@ public class Commander
     {
         private DummyMover()
         {
-            super(BigDoors.get(), null);
+            super(BigDoors.get(), null, false);
         }
 
         @Override
