+++ core/src/main/java/nl/pim16aap2/bigDoors/moveBlocks/Bridge/getNewLocation/GetNewLocation.java
@@ -4,5 +4,5 @@ import org.bukkit.Location;
 
 public interface GetNewLocation
 {
-    public Location getNewLocation(double radius, double xPos, double yPos, double zPos, int index);
+    public Location getNewLocation(double radius, double xPos, double yPos, double zPos);
 }
